fix : 카드 더 뽑는 구조 변경
