+++ src/main/kotlin/blackjack/domain/Drawable.kt
@@ -1,0 +1,5 @@
+package blackjack.domain
+
+interface Drawable {
+    fun canDraw(): Boolean
+}
