+++ src/main/kotlin/blackjack/view/InputView.kt
@@ -1,9 +1,7 @@
 package blackjack.view
 
 import blackjack.domain.Dealer
-import blackjack.domain.Participant
 import blackjack.domain.Player
-import blackjack.exception.InvalidInputValueException
 
 object InputView {
 
@@ -28,23 +26,5 @@ object InputView {
         }
     }
 
-    fun needMoreCard(participant: Participant): Boolean {
-        println()
-        println("${participant.name}는 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)")
-        return readln().toBoolean().also {
-            if (it) participant.setBlackJackStatusStay()
-        }
-    }
-
-    private fun String.toBoolean(): Boolean {
-        return when (uppercase()) {
-            YES -> true
-            NO -> false
-            else -> throw InvalidInputValueException()
-        }
-    }
-
-    private const val YES = "Y"
-    private const val NO = "N"
     private const val DEALER = "딜러"
 }
